Fix set user to store
