+++ wbooks/src/services/AuthService.ts
@@ -1,3 +1,6 @@
+import AsyncStorage from '@react-native-async-storage/async-storage';
+import Reactotron from 'reactotron-react-native';
+
 interface Props {
   email: string;
   password?: string;
@@ -9,3 +12,13 @@ export const login = ({ email }: Props) =>
     data: { headers: { uid: email, token: 'AE12OI45PK56LK', client: 'secret_variable' }, currentUser: email },
     problem: ''
   });
+
+export const getCurrentUser = async () => {
+  try {
+    const authData = await AsyncStorage.getItem('authData');
+    return authData === null ? null : JSON.parse(authData).currentUser;
+  } catch (e) {
+    if (__DEV__) Reactotron.log(e);
+    return null;
+  }
+};
